refactor(daemon): 排队器从网关 crate 移到 daemon (#933)
session_queue.rs（452 行）管的是每会话串行、队列深度、消息超时、空闲回收
——全是运行时行为，不是协议。它零依赖 gateway crate 里的任何东西，而调用
方 core/sink.rs 早就在 daemon 侧，所以搬过来是纯粹的归位。

gateway crate: 18,233 → 17,612 行。

顺带：它自带的 9 个用例第一次进入 `cargo test --bin amuxd` 的覆盖范围
（之前在 gateway crate 里，daemon 的测试跑不到），所以用例数从 1353 涨到
1362。

注：get_mcp_empty_workspace_returns_empty_map 在本机红，它读的是真实的
~/.amuxd/mcp.json（本机有内容），与本次改动无关，workspace_control.rs 整个
分支一行没动。

via [Happy](https://happy.engineering)
