+++ apps/daemon/src/channels/manager.rs
@@ -142,7 +142,7 @@ impl ChannelManager {
             // shorter and a message waiting behind a healthy long turn gets
             // dropped for being late.
             queue: Arc::new(
-                teamclu_gateway::session_queue::SessionQueue::with_message_timeout(
+                crate::channels::session_queue::SessionQueue::with_message_timeout(
                     std::time::Duration::from_secs(driver.caps().turn_timeout_secs),
                 ),
             ),
