+++ apps/daemon/src/channels/session_queue.rs
